requirements.txt
streamlit>=1.31.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
groq>=0.4.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
validators>=0.22.0
urllib3>=2.0.0
lxml>=4.9.0
